requirements.txt
flask
torch
transformers
bert-extractive-summarizer
moviepy
requests
googletrans==4.0.0-rc1